[Activity 1]
(ADDED) Exercises 5 and 6
(CHANGED) Exercises 1 and 2 now use int's instead of longs

+++ Activity1/Exercise2/Program.cs
@@ -5,18 +5,18 @@
         private static void Main()
         {
             Console.Write("Enter the first integer: ");
-            long number1 = Convert.ToInt64(Console.ReadLine());
+            int number1 = Convert.ToInt32(Console.ReadLine());
             Console.Write("Enter the second integer: ");
-            long number2 = Convert.ToInt64(Console.ReadLine());
+            int number2 = Convert.ToInt32(Console.ReadLine());
             Console.Write("Enter the third integer: ");
-            long number3 = Convert.ToInt64(Console.ReadLine());
+            int number3 = Convert.ToInt32(Console.ReadLine());
 
             Console.Write("The highest number is: {0}", GetHighestNumber(number1, number2, number3));
         }
 
-        private static long GetHighestNumber(long number1, long number2, long number3)
+        private static long GetHighestNumber(int number1, int number2, int number3)
         {
-            long[] numbers = { number1, number2, number3 };
+            int[] numbers = { number1, number2, number3 };
             return numbers.Max();
         }
     }
